# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
...
  "/user/create":
    post:
      description: Creates a new user.
      tags:
        - User
      requestBody:
        description: "A username, email and a password."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              username: "username123"
              email: "my@email.com"
              password: "newpassw0rd"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  usernameAvailable:
                    type: boolean
                    example: true
                  emailAvailable:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully created the user."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: "5f3b3b3b3b3b3b3b3b3b3b3b"
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
...
  "/user/journal/create":
    post:
      description: "Saves a journal."
      tags:
        - Journal
      requestBody:
        description: "User's document ID and the journal to save."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                title:
                  type: string
                entry:
                  type: string
                category:
                  type: string
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              title: "Journal"
              entry: "A journal about things."
              category: "Today"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Journal saved."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": [],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/user/journal/category/create":
    post:
      description: "Saves a new category."
      tags:
        - Journal
      requestBody:
        description: "A users document ID and a category string."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  description: "The user's document id."
                  type: string
                category:
                  description: "The category string."
                  type: string
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Category saved."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": ["New Category"],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/user/journal/create-many":
    post:
      description: "Saves a category on the journal(s)."
      tags:
        - Journal
      requestBody:
        description: "A users document ID, the journal document id's, and a category string."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  description: "The user's document id."
                  type: string
                journalIds:
                  description: "The journal document id(s)."
                  type: array
                  items:
                    type: string
                category:
                  description: "The category string."
                  type: string
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              journalIds:
                [
                  "eyJhbGciOifgIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "eyJhbhgiOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                ]
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Journals updated."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": ["New Category"],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...